+++ src/main/java/org/example/goormssd/usermanagementbackend/config/SecurityConfig.java
@@ -3,7 +3,7 @@ package org.example.goormssd.usermanagementbackend.config;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.goormssd.usermanagementbackend.util.JwtAuthenticationFilter;
+import org.example.goormssd.usermanagementbackend.config.security.JwtAuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
